\name{atm}
\docType{data}
\alias{atm}
\title{S&P500 Implied Volatilities - ATM}
\description{
  This dataset covers daily time-series of Black-Scholes implied volatilities from at-the-money Options on the S&P500 Equity Index, aka implied volatility term-structure.  
}
\usage{atm}
\format{A \code{data.frame} with 2845 observations and 8 different maturities. Column names indicate different maturities: e.g. T30 means ATM implied volatility for the 30-day horizon. Row names represent date of observation.}
\source{Thomson Reuter's Datastream}
\keyword{datasets}
