\name{otm}
\docType{data}
\alias{otm}
\title{S&P500 Implied Volatilities - OTM}
\description{
  This dataset covers daily time-series of Black-Scholes implied volatilities from out-of-money Options on the S&P500 Equity Index with a target maturity of 1 month.  
}
\usage{otm}
\format{A \code{data.frame} with 2845 observations and 9 different levels of moneyness. Column names indicate degrees of moneyness: e.g. M80 means Option data of 80\% moneyness. Row names represent date of observation.}
\source{Thomson Reuter's Datastream}
\keyword{datasets}

